+++ src/internal/m365/onedrive_test.go
@@ -12,6 +12,7 @@ import (
 	"github.com/stretchr/testify/require"
 	"github.com/stretchr/testify/suite"
 
+	"github.com/alcionai/corso/src/internal/common/dttm"
 	"github.com/alcionai/corso/src/internal/common/ptr"
 	"github.com/alcionai/corso/src/internal/m365/graph"
 	odConsts "github.com/alcionai/corso/src/internal/m365/onedrive/consts"
@@ -517,19 +518,17 @@ func testRestoreAndBackupMultipleFilesAndFoldersNoPermissions(
 				collectionsLatest:   expected,
 			}
 
-			rc := testdata.DefaultRestoreConfig("od_restore_and_backup_multi")
-			rc.OnCollision = control.Replace
+			restoreCfg := testdata.DefaultRestoreConfig("od_restore_and_backup_multi")
+			restoreCfg.OnCollision = control.Replace
+			restoreCfg.IncludePermissions = true
 
 			runRestoreBackupTestVersions(
 				t,
 				testData,
 				suite.Tenant(),
 				[]string{suite.ResourceOwner()},
-				control.Options{
-					RestorePermissions: true,
-					ToggleFeatures:     control.Toggles{},
-				},
-				rc)
+				control.DefaultOptions(),
+				restoreCfg)
 		})
 	}
 }
@@ -768,19 +767,17 @@ func testPermissionsRestoreAndBackup(suite oneDriveSuite, startVersion int) {
 				collectionsLatest:   expected,
 			}
 
-			rc := testdata.DefaultRestoreConfig("perms_restore_and_backup")
-			rc.OnCollision = control.Replace
+			restoreCfg := testdata.DefaultRestoreConfig("perms_restore_and_backup")
+			restoreCfg.OnCollision = control.Replace
+			restoreCfg.IncludePermissions = true
 
 			runRestoreBackupTestVersions(
 				t,
 				testData,
 				suite.Tenant(),
 				[]string{suite.ResourceOwner()},
-				control.Options{
-					RestorePermissions: true,
-					ToggleFeatures:     control.Toggles{},
-				},
-				rc)
+				control.DefaultOptions(),
+				restoreCfg)
 		})
 	}
 }
@@ -860,19 +857,17 @@ func testPermissionsBackupAndNoRestore(suite oneDriveSuite, startVersion int) {
 				collectionsLatest:   expected,
 			}
 
-			rc := testdata.DefaultRestoreConfig("perms_backup_no_restore")
-			rc.OnCollision = control.Replace
+			restoreCfg := testdata.DefaultRestoreConfig("perms_backup_no_restore")
+			restoreCfg.OnCollision = control.Replace
+			restoreCfg.IncludePermissions = true
 
 			runRestoreBackupTestVersions(
 				t,
 				testData,
 				suite.Tenant(),
 				[]string{suite.ResourceOwner()},
-				control.Options{
-					RestorePermissions: false,
-					ToggleFeatures:     control.Toggles{},
-				},
-				rc)
+				control.DefaultOptions(),
+				restoreCfg)
 		})
 	}
 }
@@ -1067,19 +1062,17 @@ func testPermissionsInheritanceRestoreAndBackup(suite oneDriveSuite, startVersio
 				collectionsLatest:   expected,
 			}
 
-			rc := testdata.DefaultRestoreConfig("perms_inherit_restore_and_backup")
-			rc.OnCollision = control.Replace
+			restoreCfg := testdata.DefaultRestoreConfig("perms_inherit_restore_and_backup")
+			restoreCfg.OnCollision = control.Replace
+			restoreCfg.IncludePermissions = true
 
 			runRestoreBackupTestVersions(
 				t,
 				testData,
 				suite.Tenant(),
 				[]string{suite.ResourceOwner()},
-				control.Options{
-					RestorePermissions: true,
-					ToggleFeatures:     control.Toggles{},
-				},
-				rc)
+				control.DefaultOptions(),
+				restoreCfg)
 		})
 	}
 }
@@ -1264,19 +1257,17 @@ func testLinkSharesInheritanceRestoreAndBackup(suite oneDriveSuite, startVersion
 				collectionsLatest:   expected,
 			}
 
-			rc := testdata.DefaultRestoreConfig("linkshares_inherit_restore_and_backup")
-			rc.OnCollision = control.Replace
+			restoreCfg := testdata.DefaultRestoreConfig("linkshares_inherit_restore_and_backup")
+			restoreCfg.OnCollision = control.Replace
+			restoreCfg.IncludePermissions = true
 
 			runRestoreBackupTestVersions(
 				t,
 				testData,
 				suite.Tenant(),
 				[]string{suite.ResourceOwner()},
-				control.Options{
-					RestorePermissions: true,
-					ToggleFeatures:     control.Toggles{},
-				},
-				rc)
+				control.DefaultOptions(),
+				restoreCfg)
 		})
 	}
 }
@@ -1383,16 +1374,16 @@ func testRestoreFolderNamedFolderRegression(
 				collectionsLatest:   expected,
 			}
 
+			restoreCfg := control.DefaultRestoreConfig(dttm.HumanReadableDriveItem)
+			restoreCfg.IncludePermissions = true
+
 			runRestoreTestWithVersion(
 				t,
 				testData,
 				suite.Tenant(),
 				[]string{suite.ResourceOwner()},
-				control.Options{
-					RestorePermissions: true,
-					ToggleFeatures:     control.Toggles{},
-				},
-			)
+				control.DefaultOptions(),
+				restoreCfg)
 		})
 	}
 }