+++ src/internal/operations/inject/inject.go
@@ -37,10 +37,7 @@ type (
 	RestoreConsumer interface {
 		ConsumeRestoreCollections(
 			ctx context.Context,
-			backupVersion int,
-			selector selectors.Selector,
-			restoreCfg control.RestoreConfig,
-			opts control.Options,
+			rcc RestoreConsumerConfig,
 			dcs []data.RestoreCollection,
 			errs *fault.Bus,
 			ctr *count.Bus,
@@ -49,6 +46,7 @@ type (
 		Wait() *data.CollectionStats
 
 		CacheItemInfoer
+		PopulateProtectedResourceIDAndNamer
 	}
 
 	CacheItemInfoer interface {
@@ -74,6 +72,25 @@ type (
 		Wait() *data.CollectionStats
 
 		CacheItemInfoer
+	}
+
+	PopulateProtectedResourceIDAndNamer interface {
+		// PopulateProtectedResourceIDAndName takes the provided owner identifier and produces
+		// the owner's name and ID from that value.  Returns an error if the owner is
+		// not recognized by the current tenant.
+		//
+		// The id-name cacher should be optional.  Some processes will look up all owners in
+		// the tenant before reaching this step.  In that case, the data gets handed
+		// down for this func to consume instead of performing further queries.  The
+		// data gets stored inside the controller instance for later re-use.
+		PopulateProtectedResourceIDAndName(
+			ctx context.Context,
+			owner string, // input value, can be either id or name
+			ins idname.Cacher,
+		) (
+			id, name string,
+			err error,
+		)
 	}
 
 	RepoMaintenancer interface {